+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Each list item should be prefixed with `(patch)` or `(minor)` or `(major)`.
 See `PUBLISH.md` for instructions on how to publish a new version.
 -->
 
+
+## v1.2.0 - e4f2907
+
 - (minor) Add syntax for expandable details
 - (patch) Dependency updates
 - (minor) Add styling for Markdown tables